feat(api): add `validator_rounds` table (#213)
* add `validator_rounds` table

* style: resolve style guide violations

* fix calculateRound

* only write validator set per round

* style: resolve style guide violations

* fix

* use isNewRound instead of upsert

* style: resolve style guide violations

* fix test

---------

+++ packages/api-database/source/models/block.ts
@@ -34,7 +34,7 @@ export class Block {
 
 	@Column({
 		nullable: false,
-		type: "integer",
+		type: "bigint",
 		unique: true,
 	})
 	public readonly height!: number;
